@@ -35,6 +35,7 @@
     "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.1",
     "babel-plugin-react-transform": "^2.0.0",
+    "babel-plugin-syntax-decorators": "^6.13.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-polyfill": "^6.3.14",
     "babel-preset-es2015": "^6.3.13",
@@ -62,7 +63,7 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
     "gapi": "0.0.3",
-    "history": "^4.4.1",
+    "history": "^4.7.2",
     "http-proxy": "^1.12.0",
     "imports-loader": "^0.6.5",
     "jasmine-core": "^2.4.1",
@@ -75,7 +76,6 @@
     "less": "^2.5.3",
     "less-loader": "^2.2.2",
     "lodash": "^4.5.1",
-    "material-ui": "^0.16.4",
     "mocha": "^3.0.2",
     "morgan": "^1.6.1",
     "node-sass": "^3.4.2",
@@ -84,21 +84,20 @@
     "q": "^1.4.1",
     "qs": "^6.1.0",
     "rc-datepicker": "^4.0.1",
-    "react": "^15.3.1",
+    "react": "^16.3.0",
     "react-addons-css-transition-group": "^15.3.1",
     "react-calendar-component": "^1.0.0",
     "react-date-picker": "^5.3.28",
     "react-datepicker": "^0.37.0",
     "react-document-meta": "^2.0.0-rc2",
-    "react-dom": "^15.1.0",
+    "react-dom": "^16.3.0",
     "react-forms": "^2.0.0-beta33",
     "react-hot-loader": "^1.3.0",
     "react-loading-order-with-animation": "^1.0.0",
     "react-onclickoutside": "^5.3.3",
     "react-redux": "^4.3.0",
-    "react-router": "3.0.0",
+    "react-router": "^3.2.1",
     "react-router-redux": "^4.0.0",
-    "react-tap-event-plugin": "^2.0.1",
     "react-transform-hmr": "^1.0.1",
     "redux": "^3.2.1",
     "redux-form": "^6.0.1",
@@ -117,7 +116,12 @@
     "yargs": "^6.5.0"
   },
   "dependencies": {
+    "@material-ui/codemod": "^1.1.0",
+    "@material-ui/core": "^1.1.0",
+    "@material-ui/icons": "^1.1.0",
+    "material-ui": "^0.20.1",
+    "prop-types": "^15.6.0",
     "react-helmet": "^5.2.0",
-    "redux-form-material-ui": "^4.3.4"
+    "react-tap-event-plugin": "^3.0.2"
   }
 }
