@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import * as actionCreators from '../../../actions/auth';
@@ -79,7 +80,7 @@ export default class AdminView extends React.Component {
 
 
 AdminView.propType={
-    userName:React.PropTypes.string,
+    userName: PropTypes.string,
 };
 
 
