@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import * as actionCreators from '../../../actions/auth';
@@ -116,7 +117,7 @@ export default class StaffPwd extends React.Component {
 
 
 StaffPwd.propType={
-    userName:React.PropTypes.string,
+    userName: PropTypes.string,
 
 };
 
