@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { browserHistory } from 'react-router';
@@ -80,8 +81,8 @@ export function requireAuthentication(Component) {
     }
 
     AuthenticatedComponent.propTypes = {
-        loginUserSuccess: React.PropTypes.func,
-        isAuthenticated: React.PropTypes.bool,
+        loginUserSuccess: PropTypes.func,
+        isAuthenticated: PropTypes.bool,
     };
 
     return connect(mapStateToProps, mapDispatchToProps)(AuthenticatedComponent);
