@@ -1,5 +1,5 @@
 """
-Django settings for InfoSys project.
+Django settings for top project.
 
 Generated by 'django-admin startproject' using Django 2.0.
 
@@ -35,6 +35,7 @@ ALLOWED_HOSTS = []
 INSTALLED_APPS = [
     # 'LogIn.apps.LoginConfig',
     'authentication.apps.AuthenticationConfig',
+
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -55,7 +56,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'InfoSys.urls'
+ROOT_URLCONF = 'top.urls'
 
 WEBPACK_LOADER = {
     'DEFAULT': {
@@ -64,16 +65,21 @@ WEBPACK_LOADER = {
     }
 }
 
+# TODO
+'''
 REST_FRAMEWORK = {
+    'PAGE_SIZE': 10,
+
+    # 设置所有接口都需要被验证
     'DEFAULT_PERMISSION_CLASSES': (
         'rest_framework.permissions.IsAuthenticated',
     ),
     'DEFAULT_AUTHENTICATION_CLASSES': (
         'rest_framework_jwt.authentication.JSONWebTokenAuthentication',
-        'rest_framework.authentication.SessionAuthentication',
-        'rest_framework.authentication.BasicAuthentication',
+        
     ),
 }
+'''
 
 
 TEMPLATES = [
@@ -92,7 +98,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'InfoSys.wsgi.application'
+WSGI_APPLICATION = 'top.wsgi.application'
 
 # Database
 # https://docs.djangoproject.com/en/2.0/ref/settings/#databases
@@ -164,7 +170,7 @@ JWT_AUTH = {
 
 LANGUAGE_CODE = 'en-us'
 
-TIME_ZONE = 'Asia/Kolkata'
+TIME_ZONE = 'UTC'
 
 USE_I18N = True
 
