@@ -1,4 +1,4 @@
-"""InfoSys URL Configuration
+"""top URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.0/topics/http/urls/
