@@ -11,33 +11,20 @@ import {
 import {parseJSON} from '../utils/misc';
 import {
     get_token,
-    create_user,
-    data_about_user
 } from '../utils/http_functions';
-import {GET_USER_SUCCESS} from "../constants";
 
-
-export function loginUserSuccess(token) {
+export function loginUserSuccess(token, type, status) {
     localStorage.setItem('token', token);
+    localStorage.setItem('type', type);
     return {
         type: LOGIN_USER_SUCCESS,
         payload: {
             token,
+            user_type: type,
+            status: status,
         },
     };
 }
-
-
-export function getUserSuccess(response) {
-    localStorage.setItem('userName', response.name);
-    return {
-        type: GET_USER_SUCCESS,
-        payload: {
-            userName: response.name,
-        },
-    };
-}
-
 
 export function loginUserFailure(error) {
     localStorage.removeItem('token');
@@ -49,6 +36,7 @@ export function loginUserFailure(error) {
         },
     };
 }
+
 
 export function loginUserRequest() {
     return {
@@ -84,7 +72,6 @@ export function loginUser(username, password, type) {
             .then(parseJSON)
             .then(response => {
                 try {
-                    dispatch(loginUserSuccess(response.token));
                     fetch('/api/user',{
                         method: 'GET',
                         headers: {
@@ -94,33 +81,33 @@ export function loginUser(username, password, type) {
                     })
                         .then((response) => response.json())
                         .then((data) => {
-                            var json =  JSON.parse(data);//转换为json对象
+                            var json =  JSON.parse(data);
                             var user_type = json[0].user_type;
-                            console.log(user_type);
-                            console.log(USER[user_type]);
-                            if (USER[type] != user_type) {
+                            if (USER[type] !== user_type) {
                                 dispatch(loginUserFailure({
                                     response: {
                                         status: 403,
-                                        statusText: 'Type Error',
+                                        statusText: '用户类型错误',
                                     },
                                 }));
                             } else {
-                                dispatch(getUserSuccess(json[0]));
+                                var id_number = json[0].id_number;
+                                var cut_id_number = id_number.substr(id_number.length-6);
+                                if (cut_id_number === password) {
+                                    dispatch(loginUserSuccess(response.token, user_type, 200));
+                                } else {
+                                    dispatch(loginUserSuccess(response.token, user_type, 201));
+                                }
                                 browserHistory.push('/main');
-                                // if (user_type == "1") {
-                                //     browserHistory.push('/student');
-                                // } else if (user_type == "2") {
-                                //     browserHistory.push('/teacher');
-                                // } else if (user_type == "3") {
-                                //     browserHistory.push('/staff');
-                                // } else if (user_type == "4") {
-                                //     browserHistory.push('/admin');
-                                // }
                             }
                         })
                         .catch((e) => {
-                            alert("身份验证失效，请重新登录");
+                            dispatch(loginUserFailure({
+                                response: {
+                                    status: 500,
+                                    statusText: '获取信息失败',
+                                },
+                            }));
                             browserHistory.push("/login");
                         });
                 } catch (e) {
@@ -128,7 +115,7 @@ export function loginUser(username, password, type) {
                     dispatch(loginUserFailure({
                         response: {
                             status: 403,
-                            statusText: 'Invalid token',
+                            statusText: '身份验证失效',
                         },
                     }));
                 }
@@ -137,12 +124,11 @@ export function loginUser(username, password, type) {
                 dispatch(loginUserFailure({
                     response: {
                         status: 403,
-                        statusText: 'Invalid username or password',
+                        statusText: '用户名或密码错误',
                     },
                 }));
             });
-
-
     };
+    return true;
 
 }
