@@ -1,20 +1,23 @@
 import React from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import classNames from 'classnames';
+import {bindActionCreators} from 'redux';
+import {withStyles} from '@material-ui/core/styles';
+import {connect} from 'react-redux';
 import * as actionCreators from '../../../actions/auth';
-import Paper from 'material-ui/Paper';
-import RaisedButton from 'material-ui/RaisedButton';
+import Paper from '@material-ui/core/Paper';
+import Button from '@material-ui/core/Button';
 import {browserHistory} from "react-router";
-import Divider from 'material-ui/Divider';
-import TextField from 'material-ui/TextField';
-import Dialog from 'material-ui/Dialog';
-import { Card, CardHeader, CardText } from 'material-ui/Card';
-import AppBar from 'material-ui/AppBar';
-import MenuItem from 'material-ui/MenuItem';
-import Drawer from 'material-ui/Drawer';
-import FlatButton from 'material-ui/FlatButton';
-import ContentCopy from 'material-ui/svg-icons/content/content-copy';
+import Divider from '@material-ui/core/Divider';
+import TextField from '@material-ui/core/TextField';
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
+import Card from '@material-ui/core/Card';
 import StaffBar from "./StaffBar";
+import MenuItem from '@material-ui/core/MenuItem';
 
 function mapStateToProps(state) {
     return {
@@ -27,35 +30,105 @@ function mapDispatchToProps(dispatch) {
     return bindActionCreators(actionCreators, dispatch);
 }
 
-const style = {
-    marginLeft: 20,
+const drawerWidth = 240;
+const styles = theme => ({
+    root: {
+        flexGrow: 1,
+    },
+    appFrame: {
+        height: '90%',
+        zIndex: 1,
+        overflow: 'hidden',
+        position: 'relative',
+        display: 'flex',
+        width: '100%',
+    },
+    drawerHeader: {
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'flex-end',
+        padding: '0 8px',
+        ...theme.mixins.toolbar,
+    },
+    content: {
+        flexGrow: 1,
+        backgroundColor: theme.palette.background.default,
+        padding: theme.spacing.unit * 3,
+        transition: theme.transitions.create('margin', {
+            easing: theme.transitions.easing.sharp,
+            duration: theme.transitions.duration.leavingScreen,
+        }),
+    },
+    'content-left': {
+        marginLeft: -drawerWidth,
+    },
+    contentShift: {
+        transition: theme.transitions.create('margin', {
+            easing: theme.transitions.easing.easeOut,
+            duration: theme.transitions.duration.enteringScreen,
+        }),
+    },
+    'contentShift-left': {
+        marginLeft: 0,
+    },
+    Button: {
+        marginTop: 50,
+        marginLeft: '45%',
+        marginBottom: 20,
+    },
+    Paper: {
+        marginLeft: 100,
+        marginRight: 100,
+        marginTop: 50
+    },
+    TextField: {
+        width: '90%',
+        marginLeft: '5%',
+    }
+});
 
-};
+
+const ranges = [
+    {
+        value: '0',
+        label: '公共通识课',
+    },
+    {
+        value: '1',
+        label: '专业选修课',
+    },
+    {
+        value: '2',
+        label: '专业必修课',
+    },
+];
 
 
 @connect(mapStateToProps, mapDispatchToProps)
-export default class CreateLesson extends React.Component {
+class CreateLesson extends React.Component {
 
-    constructor(props){
+    constructor(props) {
         super(props);
 
-        this.state={
-            userName : '',
-            course_id : '',
-            course_name : '',
-            course_credit : -1,
-            course_cap : -1,
-            course_room : '',
-            course_assessment:'',
-            id_error : '',
-            name_error : '',
-            credit_error : '',
-            cap_error : '',
-            room_error : '',
-            disabled : true,
-            open: false,
-            dialog_text:'',
-            drawerOpen:true,
+        this.state = {
+            userName: '',
+            course_id: '',
+            course_name: '',
+            course_credit: -1,
+            course_cap: -1,
+            course_room: '',
+            course_assessment: '',
+            course_type: '',
+            id_error: '',
+            name_error: '',
+            credit_error: '',
+            cap_error: '',
+            room_error: '',
+            disabled: true,
+            dialogOpen: false,
+            dialog_text: '',
+            open: true,
+            anchor: 'left',
         };
     }
 
@@ -64,10 +137,10 @@ export default class CreateLesson extends React.Component {
     }
 
     handleClose = () => {
-        this.setState({open: false});
+        this.setState({dialogOpen: false});
     };
 
-    isDisabled(){
+    isDisabled() {
         let name_is_valid = false;
         let id_is_valid = false;
         let credit_is_valid = false;
@@ -76,7 +149,7 @@ export default class CreateLesson extends React.Component {
 
         if (this.state.course_id === '') {
             this.setState({
-                id_error: "course ID is required",
+                id_error: "课程号为必填项",
             });
         } else {
             if (this.state.course_id.length < 10) {
@@ -86,14 +159,14 @@ export default class CreateLesson extends React.Component {
                 });
             } else {
                 this.setState({
-                    id_error: "the course ID is too long",
+                    id_error: "课程号过长",
                 });
             }
         }
 
         if (this.state.course_name === '') {
             this.setState({
-                name_error: "Course name is required.",
+                name_error: "课程名为必填项",
             });
         } else {
             if (this.state.course_name.length < 40) {
@@ -103,15 +176,15 @@ export default class CreateLesson extends React.Component {
                 });
             } else {
                 this.setState({
-                    email_error_text: 'the course name is too long',
+                    email_error_text: '课程名过长',
                 });
             }
         }
 
 
-        if (this.state.course_credit === -1 ) {
+        if (this.state.course_credit === -1) {
             this.setState({
-                credit_error: 'course credit is required',
+                credit_error: '课程学分为必填项',
             });
         } else {
             if (this.state.course_credit <= 10 && this.state.course_credit > 0) {
@@ -121,15 +194,15 @@ export default class CreateLesson extends React.Component {
                 });
             } else {
                 this.setState({
-                    password_error_text: 'invalid credit',
+                    password_error_text: '课程学分非法',
                 });
             }
         }
 
 
-        if (this.state.course_cap === -1 ) {
+        if (this.state.course_cap === -1) {
             this.setState({
-                cap_error: 'Course capacity is required',
+                cap_error: '课程容量为必填项',
             });
         } else {
             if (this.state.course_cap <= 300 && this.state.course_cap > 0) {
@@ -139,15 +212,15 @@ export default class CreateLesson extends React.Component {
                 });
             } else {
                 this.setState({
-                    cap_error: 'invalid capacity',
+                    cap_error: '课程容量非法',
                 });
             }
         }
 
 
-        if (this.state.course_room === '' ) {
+        if (this.state.course_room === '') {
             this.setState({
-                room_error: 'classroom is required',
+                room_error: '教室为必填项',
             });
         } else {
             room_is_valid = true;
@@ -156,7 +229,7 @@ export default class CreateLesson extends React.Component {
             });
         }
 
-        if (id_is_valid && name_is_valid && credit_is_valid && room_is_valid && cap_is_valid ) {
+        if (id_is_valid && name_is_valid && credit_is_valid && room_is_valid && cap_is_valid) {
             this.setState({
                 disabled: false,
             });
@@ -165,29 +238,28 @@ export default class CreateLesson extends React.Component {
     }
 
 
-    handleKeyPress(e){
+    handleKeyPress(e) {
         this.isDisabled();
-        this.state.open = true;
+        this.state.dialogOpen = true;
         if (!this.state.disabled) {
             var status = this.apply(e).status;
-            if (status == 200) {
+            if (status === 200) {
                 this.setState({
-                dialog_text: "new Lesson added successfully",
-            });
+                    dialog_text: "课程添加成功",
+                });
             } else {
                 this.setState({
-                    dialog_text: "new Lesson failed to add"});
+                    dialog_text: "课程添加失败"
+                });
             }
-
         }
-        else{
+        else {
             this.setState({
-                dialog_text: "invalid submit",
+                dialog_text: "课程信息非法",
             });
-
         }
-
     }
+
     apply(e) {
         e.preventDefault();
         let formData = new FormData();
@@ -215,101 +287,130 @@ export default class CreateLesson extends React.Component {
             });
     }
 
-    changeValue(e, type) {
-        const value = e.target.value;
-        const next_state = {};
-        next_state[type] = value;
-        this.setState(next_state);
-    }
+    handleChange = prop => event => {
+        this.setState({[prop]: event.target.value});
+    };
 
-    handleClick(){
-        this.setState({drawerOpen: !this.state.drawerOpen});
+    handleClick() {
+        this.setState({open: !this.state.open});
     }
 
     render() {
-        const actions = [
-            <FlatButton
-                label="Close"
-                primary={true}
-                onClick={this.handleClose}
-            />,
-        ];
-
-        const contentStyle = {
-            transition: 'margin-left 450ms cubic-bezier(0.23, 1, 0.32, 1)' ,
-            margin: '50'
-        };
-        if (this.state.drawerOpen) {
-            contentStyle.marginLeft = 220;
-        }
-        else{
-            contentStyle.marginLeft = 50;
-        }
+        const {classes, theme} = this.props;
+        const {anchor, open} = this.state;
 
         return (
-
-            <div >
-                <StaffBar handleClick={this.handleClick.bind(this)} drawerOpen={this.state.drawerOpen}/>
-                {/*<Drawer open={this.state.drawerOpen} width={200} >*/}
-                    {/*<div >*/}
-                        {/*<AppBar onLeftIconButtonTouchTap={() => this.setState({ drawerOpen: !this.state.drawerOpen })}/>*/}
-                        {/*<MenuItem primaryText="个人信息" leftIcon={<ContentCopy />}  />*/}
-                        {/*<MenuItem primaryText="课程信息" leftIcon={<ContentCopy />} onClick={(e) => this.dispatchNewRoute(e, '/staff/lessons')}/>*/}
-                        {/*<MenuItem primaryText="添加课程" leftIcon={<ContentCopy />} onClick={(e) => this.dispatchNewRoute(e, '/staff/createLesson')}/>*/}
-                        {/*<MenuItem primaryText="删除课程" leftIcon={<ContentCopy />} onClick={(e) => this.dispatchNewRoute(e, '/staff/delete')}/>*/}
-                        {/*<MenuItem primaryText="处理申请课程" leftIcon={<ContentCopy />} onClick={(e) => this.dispatchNewRoute(e, '/staff/process')} />*/}
-                    {/*</div>*/}
-
-                {/*</Drawer>*/}
-
-                {/*<AppBar title={'Hello, '+this.state.userName}*/}
-                        {/*onLeftIconButtonTouchTap={() => this.setState({ drawerOpen: !this.state.drawerOpen })}*/}
-                        {/*iconElementRight={*/}
-                            {/*<FlatButton label="Back"  onClick={() => browserHistory.push("/staff")}/>*/}
-                        {/*}*/}
-                {/*/>*/}
-
-                <Card style={contentStyle}>
-                    <CardHeader title="新建课程信息" />
-
-                    <Paper zDepth={2} style={{marginLeft:100,marginRight:100, marginTop:50}}>
-                        <TextField hintText="Course ID" errorText={this.state.id_error} style={style} underlineShow={false} onChange={(e) => this.changeValue(e, 'course_id')}/>
-                        <Divider />
-                        <TextField hintText="Course Name" errorText={this.state.name_error} style={style} underlineShow={false} onChange={(e) => this.changeValue(e, 'course_name')}/>
-                        <Divider />
-                        <TextField hintText="Credit" errorText={this.state.credit_error} style={style} underlineShow={false} onChange={(e) => this.changeValue(e, 'course_credit')}/>
-                        <Divider />
-                        <TextField hintText="Capacity" errorText={this.state.cap_error} style={style} underlineShow={false} onChange={(e) => this.changeValue(e, 'course_cap')}/>
-                        <Divider />
-                        <TextField hintText="Classroom" errorText={this.state.room_error} style={style} underlineShow={false} onChange={(e) => this.changeValue(e, 'course_room')}/>
-                        <Divider />
-                        <TextField hintText="Assessment (Optional)"  style={style} underlineShow={false} onChange={(e) => this.changeValue(e, 'course_assessment')}/>
-                        <Divider />
-                    </Paper>
-                    <RaisedButton
-                        style={{ marginTop: 50 , marginLeft:'45%',marginBottom:20}}
-                        label="Submit"
-                        onClick={(e) => this.handleKeyPress(e)}
-                    />
-
-                </Card>
+            <div>
+                <div className={classes.root}>
+                    <div className={classes.appFrame}>
+                        <StaffBar click={this.handleClick.bind(this)} open={this.state.open}/>
+                        <Card
+                            className={classNames(classes.content, classes[`content-${anchor}`], {
+                                [classes.contentShift]: open,
+                                [classes[`contentShift-${anchor}`]]: open,
+                            })}
+                        >
+                            <div className={classes.drawerHeader}/>
+                            <Paper>
+                                <TextField
+                                    className={classes.TextField}
+                                    label="课程号"
+                                    onChange={this.handleChange('course_id')}
+                                    margin="normal"
+                                    error={this.state.id_error}
+                                />
+                                <TextField
+                                    className={classes.TextField}
+                                    label="课程名"
+                                    onChange={this.handleChange('course_name')}
+                                    margin="normal"
+                                    error={this.state.name_error}
+                                />
+                                <TextField
+                                    select='true'
+                                    label="课程类型"
+                                    margin="normal"
+                                    className={classes.TextField}
+                                    value={this.state.course_type}
+                                    onChange={this.handleChange('course_type')}
+                                >
+                                    {ranges.map(option => (
+                                        <MenuItem key={option.value} value={option.value}>
+                                            {option.label}
+                                        </MenuItem>
+                                    ))}
+                                </TextField>
+                                <TextField
+                                    className={classes.TextField}
+                                    label="学分"
+                                    onChange={this.handleChange('course_credit')}
+                                    margin="normal"
+                                    error={this.state.credit_error}
+                                    underlineShow={false}
+                                />
+                                <TextField
+                                    className={classes.TextField}
+                                    label="容量"
+                                    onChange={this.handleChange('course_cap')}
+                                    margin="normal"
+                                    error={this.state.cap_error}
+                                />
+                                <TextField
+                                    className={classes.TextField}
+                                    label="教室"
+                                    onChange={this.handleChange('course_room')}
+                                    margin="normal"
+                                    error={this.state.room_error}
+                                />
+                                <TextField
+                                    className={classes.TextField}
+                                    label="考核方式"
+                                    onChange={this.handleChange('course_assessment')}
+                                    margin="normal"
+                                />
+                                <Divider/>
+                            </Paper>
+                            <Button
+                                variant="raised"
+                                color="primary"
+                                className={classes.Button}
+                                onClick={(e) => this.handleKeyPress(e)}
+                            >
+                                注册
+                            </Button>
+                        </Card>
+                    </div>
+                </div>
                 <Dialog
-                    title={this.state.dialog_text}
-                    actions={actions}
-                    modal={false}
-                    open={this.state.open}
-                    onRequestClose={this.handleClose}
+                    open={this.state.dialogOpen}
+                    onClose={this.handleClose}
+                    aria-labelledby="alert-dialog-title"
+                    aria-describedby="alert-dialog-description"
                 >
-
+                    <DialogTitle id="alert-dialog-title">{"提示"}</DialogTitle>
+                    <DialogContent>
+                        <DialogContentText id="alert-dialog-description">
+                            {this.state.dialog_text}
+                        </DialogContentText>
+                    </DialogContent>
+                    <DialogActions>
+                        <Button onClick={this.handleClose} color="primary">
+                            关闭
+                        </Button>
+                    </DialogActions>
                 </Dialog>
             </div>
         );
+
     }
 }
 
 
-CreateLesson.propType={
-
+CreateLesson.propType = {
+    userName: PropTypes.string,
+    classes: PropTypes.object.isRequired,
+    theme: PropTypes.object.isRequired,
 };
+export default withStyles(styles, {withTheme: true})(CreateLesson);
 
 
