@@ -8,7 +8,7 @@ export const Footer = () =>
         <div className="container">
             <div className="row">
                 <div className="col-xs-12 col-sm-12 col-md-12 col-lg-12">
-                    <p>© DZT 2016</p>
+                    <p>© InfoSys 2018</p>
                 </div>
             </div>
         </div>
