@@ -1,23 +1,29 @@
 import React from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import {withStyles} from "@material-ui/core/styles/index";
+import {bindActionCreators} from 'redux';
+import {connect} from 'react-redux';
 import * as actionCreators from '../actions/auth';
-import Paper from 'material-ui/Paper';
-import RaisedButton from 'material-ui/RaisedButton';
 import {browserHistory} from "react-router";
-import FlatButton from 'material-ui/FlatButton';
-import Drawer from 'material-ui/Drawer';
-import { Card, CardHeader, CardText , CardActions,} from 'material-ui/Card';
-import ContentCopy from 'material-ui/svg-icons/content/content-copy';
-import AppBar from 'material-ui/AppBar';
-import MenuItem from 'material-ui/MenuItem';
+import Button from '@material-ui/core/Button'
+import Card from '@material-ui/core/Card';
+import CardActions from '@material-ui/core/CardActions';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import IconButton from '@material-ui/core/IconButton';
+import MenuIcon from '@material-ui/icons/Menu';
 import {Helmet} from "react-helmet";
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
 
 
 function mapStateToProps(state) {
     return {
-        isRegistering: state.auth.isRegistering,
-        registerStatusText: state.auth.registerStatusText,
+        status: state.auth.status,
     };
 }
 
@@ -25,68 +31,79 @@ function mapDispatchToProps(dispatch) {
     return bindActionCreators(actionCreators, dispatch);
 }
 
-const style = {
-    height: 180,
-    width: 280,
-    margin: "3%",
-    textAlign: 'center',
-    display: 'inline-block',
-};
-const buttonStyle = {
-    width:200,
-    marginTop: "30%",
-    marginLeft:"13%",
-    textAlign: 'center',
-    fontSize: 20,
-    color:"#2196F3",
+const styles = theme => ({
+    Helmet: {
+        background: '#EEEEEE',
+    },
+    Typography: {
+        flex: 1,
+    },
+    Card: {
+        width: "60%",
+        marginLeft: "20%",
+        marginTop: "5%",
+    },
+    SubCard: {
+        width: '28%',
+        height: 200,
+        marginLeft: '4%',
+        marginTop: '5%',
+        marginBottom: '5%',
+        textAlign: 'center',
+        display: 'inline-block',
+    },
+    Button: {
+        marginTop: "25%",
+        marginLeft: '25%',
+        alignItems: 'center',
+        fontSize: 18,
+        color: "#2196F3",
+    }
 
-}
+
+});
+
+
 @connect(mapStateToProps, mapDispatchToProps)
-export default class MainMenu extends React.Component {
+class MainMenu extends React.Component {
 
-    constructor(props){
+    constructor(props) {
         super(props);
-        this.state={
-            username:'',
+        this.state = {
+            username: '',
             redirectTo: '',
             user_type: -1,
-            drawerOpen:true,
+            open: false,
+            dialog_text: '',
         };
     }
 
+    handleClickOpen = () => {
+        this.setState({dialog_text: "当前的密码为默认密码（身份证后6位），请及时修改密码确保账户安全", open: true});
+    };
+
+
+    handleClose = () => {
+        this.setState({open: false});
+    };
+
     componentDidMount() {
-        fetch('/api/user',{
-            method: 'GET',
-            headers: {
-                'Authorization': 'JWT '+ localStorage.getItem('token'),
-                'Content-Type': 'application/json'
-            },
-        })
-            .then((response) => response.json())
-            .then((data) => {
-                var json =  JSON.parse(data);//转换为json对象
-                var user_type = json[0].user_type;
-                this.setState({ user_type:user_type });
-                console.log("get user_type:"+this.state.user_type);
-            })
-            .catch((e) => {
-                alert("身份验证失效，请重新登录");
-                browserHistory.push("/login");
-            });
+        this.setState({user_type: localStorage.getItem('type')});
         this.setState({userName: localStorage.getItem('userName')});
+        this.handleClickOpen();
     }
 
 
     dispatchNewRoute(e) {
         e.preventDefault();
         var route = '';
-        if(this.state.user_type === 4)
+        if (this.state.user_type === '4')
             route = '/admin';
-        else if(this.state.user_type === 1)
+        else if (this.state.user_type === '1')
             route = '/student';
-        else if(this.state.user_type === 2)
+        else if (this.state.user_type === '2')
             route = '/teacher';
-        else if(this.state.user_type === 3)
+        else if (this.state.user_type === '3')
             route = '/staff';
         browserHistory.push(route);
         this.setState({
@@ -97,52 +114,86 @@ export default class MainMenu extends React.Component {
     logout(e) {
         e.preventDefault();
         this.props.logoutAndRedirect();
-
     }
 
     render() {
-
+        const {classes} = this.props;
         return (
-            <div >
+            <div>
                 <Helmet bodyAttributes={{style: 'background-color : #EEEEEE'}}/>
 
-                <AppBar title={'Hello, '+this.state.userName}
-                        onLeftIconButtonTouchTap={() => this.setState({ drawerOpen: !this.state.drawerOpen })}
-                        iconElementRight={
-                            <FlatButton label="Exit"  onClick={(e) =>this.logout(e)}/>
-                        }
-                />
-                <Card style={{width:"72%", marginLeft:"13%", marginTop:"3%"}}>
+                <AppBar position="static">
+                    <Toolbar>
+                        <IconButton color="inherit" aria-label="Menu">
+                            <MenuIcon/>
+                        </IconButton>
+                        <Typography variant="title" color="inherit" className={classes.Typography}>
+                            InfoSys
+                        </Typography>
+                        <Button color="inherit" onClick={(e) => this.logout(e)}>退出</Button>
+                    </Toolbar>
+                </AppBar>
+                <Card className={classes.Card}>
 
-                    <Card style={style} onClick={(e)=>this.dispatchNewRoute(e)}>
-                       <CardHeader title={"基础信息管理系统"} titleStyle={buttonStyle}/>
+                    <Card className={classes.SubCard}>
+                        <CardActions>
+                            <Button variant='flat' className={classes.Button}
+                                onClick={(e) => this.dispatchNewRoute(e)}>基础信息系统</Button>
+                        </CardActions>
                     </Card>
-                    <Card style={style}>
-                        <CardHeader title={"自动排课系统"} titleStyle={buttonStyle}/>
+                    <Card className={classes.SubCard}>
+                        <CardActions>
+                            <Button variant='flat' className={classes.Button}>自助排课系统</Button>
+                        </CardActions>
                     </Card>
-                    <Card style={style}>
-                        <CardHeader title={"选课系统"} titleStyle={buttonStyle}/>
+                    <Card className={classes.SubCard}>
+                        <CardActions>
+                            <Button variant='flat' className={classes.Button}>学生选课系统</Button>
+                        </CardActions>
                     </Card>
-                    <Card style={style}>
-                        <CardHeader title={"论坛交流系统"} titleStyle={buttonStyle}/>
+                    <Card className={classes.SubCard}>
+                        <CardActions>
+                            <Button variant='flat' className={classes.Button}>论坛交流系统</Button>
+                        </CardActions>
                     </Card>
-                    <Card style={style}>
-                        <CardHeader title={"在线测试系统"} titleStyle={buttonStyle}/>
+                    <Card className={classes.SubCard}>
+                        <CardActions>
+                            <Button variant='flat' className={classes.Button}>在线测试系统</Button>
+                        </CardActions>
                     </Card>
-                    <Card style={style}>
-                        <CardHeader title={"成绩管理系统"} titleStyle={buttonStyle}/>
+                    <Card className={classes.SubCard}>
+                        <CardActions>
+                            <Button variant='flat' className={classes.Button}>成绩管理系统</Button>
+                        </CardActions>
                     </Card>
                 </Card>
-
+                <Dialog
+                    open={this.state.open}
+                    onClose={this.handleClose}
+                    aria-labelledby="alert-dialog-title"
+                    aria-describedby="alert-dialog-description"
+                >
+                    <DialogTitle id="alert-dialog-title">{"安全提示"}</DialogTitle>
+                    <DialogContent>
+                        <DialogContentText id="alert-dialog-description">
+                            {this.state.dialog_text}
+                        </DialogContentText>
+                    </DialogContent>
+                    <DialogActions>
+                        <Button onClick={this.handleClose} color="primary">
+                            关闭
+                        </Button>
+                    </DialogActions>
+                </Dialog>
             </div>
         );
     }
 }
 
 
-MainMenu.propType={
-
-
+MainMenu.propType = {
+    classes: PropTypes.object.isRequired,
 };
 
+export default withStyles(styles, {withTheme: true})(MainMenu);
 
